infra/requirements.txt
-e .
aws-cdk.core==1.122.0
aws-cdk.aws_lambda==1.122.0
aws-cdk.aws_events==1.122.0
aws-cdk.aws_events_targets==1.122.0
aws-cdk.aws_iam==1.122.0
urllib3==1.26.7
datetime==4.3
aws-cdk.aws_cloudwatch==1.122.0
boto3==1.18.64
aws-cdk.aws_sns==1.122.0
aws-cdk.aws_cloudwatch_actions==1.122.0
aws_cdk.aws_dynamodb==1.122.0
aws-cdk.aws_lambda_event_sources==1.122.0
